Render mesh3d

+++ Mesh.cpp
@@ -14,7 +14,7 @@ void Mesh3D::read(std::string filename)
 	subsets.resize(header.numSubsets);
 
 	fread(vertices.data(), sizeof(Vertex), header.numVerts, fp);
-	fread(indices.data(), sizeof(unsigned int), header.numIndices, fp);
+	fread(indices.data(), sizeof(uint16_t), header.numIndices, fp);
 	fread(subsets.data(), sizeof(Subset), header.numSubsets, fp);
 
 	fclose(fp);
